refactoring blog post
