+++ src/elements/block/core-gallery/core-gallery.tsx
