+++ src/elements/block/core-image/core-image.spec.tsx
