+++ src/elements/block/core-image/core-image.tsx
