+++ src/elements/block/core-list/core-list.tsx
