+++ src/elements/block/core-paragraph/core-paragraph.tsx
