--- src/elements/core-gallery/core-gallery.docs.mdx
@@ -1,68 +1,0 @@
----
-description: 'A CoreGallery component.'
-labels: ['label1', 'label2', 'label3']
----
-
-import { CoreGallery } from './core-gallery';
-
-## React Component for rendering a gallery
-
-A gallery component
-
-### Component usage
-```js
-<CoreGallery {...{
-    name: 'core/gallery',
-    attributes: {
-      align: '',
-      images: [
-        {
-          fullUrl:
-            'http://54.204.85.23/wp-content/uploads/2021/03/1614778286060-2-scaled.jpg',
-          alt: 'man in woods'
-        },
-        {
-          fullUrl:
-            'http://54.204.85.23/wp-content/uploads/2021/03/1542118249535-2-scaled.jpg',
-          alt: 'Jesus in boat'
-        },
-        {
-          fullUrl:
-            'http://54.204.85.23/wp-content/uploads/2021/03/1615294462458-3-scaled.jpg',
-          alt: 'Jesus with 2 others'
-        }
-      ]
-    }
-  }
-} />
-```
-
-### Using props to customize the text
-
-Modify the text to see it change live:
-```js live
-<CoreGallery {...  {
-    name: 'core/gallery',
-    attributes: {
-      align: '',
-      images: [
-        {
-          fullUrl:
-            'http://54.204.85.23/wp-content/uploads/2021/03/1614778286060-2-scaled.jpg',
-          alt: 'man in woods'
-        },
-        {
-          fullUrl:
-            'http://54.204.85.23/wp-content/uploads/2021/03/1542118249535-2-scaled.jpg',
-          alt: 'Jesus in boat'
-        },
-        {
-          fullUrl:
-            'http://54.204.85.23/wp-content/uploads/2021/03/1615294462458-3-scaled.jpg',
-          alt: 'Jesus with 2 others'
-        }
-      ]
-    }
-  }
-} />
-```
