--- src/elements/core-heading/core-heading.docs.mdx
@@ -1,36 +1,0 @@
----
-description: 'A CoreHeading component.'
-labels: ['label1', 'label2', 'label3']
----
-
-import { CoreHeading } from './core-heading';
-
-## Render Headings
-
-Render an h2 heading
-
-### Component usage
-```js
-<CoreHeading {...{
-    name: 'core/heading',
-    attributes: {
-      align: 'left',
-      content: 'The parables from Matthew',
-      level: 2
-    }
-  }} />
-```
-
-### Using props to customize the text
-
-Modify the text to see it change live:
-```js live
-<CoreHeading {...{
-    name: 'core/heading',
-    attributes: {
-      align: 'left',
-      content: 'The parables from Matthew',
-      level: 2
-    }
-  }} />
-```
