--- src/elements/core-image/core-image.composition.tsx
@@ -1,16 +1,0 @@
-import React from 'react'
-import { CoreImage } from './core-image'
-
-export const BasicCoreImage = () => (
-  <CoreImage
-    {...{
-      name: 'core/image',
-      attributes: {
-        id: 23,
-        title: '',
-        url: 'http://54.204.85.23/wp-content/uploads/2020/12/1614778286060.jpg',
-        alt: ''
-      }
-    }}
-  />
-)
