--- src/elements/core-image/core-image.docs.mdx
@@ -1,38 +1,0 @@
----
-description: 'A CoreImage component.'
-labels: ['label1', 'label2', 'label3']
----
-
-import { CoreImage } from './core-image';
-
-## React Component for rendering text
-
-A basic div that renders some text
-
-### Component usage
-```js
-<CoreImage {...{
-      name: 'core/image',
-      attributes: {
-        id: 23,
-        title: 'an image',
-        url: 'http://54.204.85.23/wp-content/uploads/2020/12/1614778286060.jpg',
-        alt: 'yoyo'
-      }
-    }} />
-```
-
-### Using props to customize the text
-
-Modify the text to see it change live:
-```js live
-<CoreImage {...{
-      name: 'core/image',
-      attributes: {
-        id: 23,
-        title: 'an image',
-        url: 'http://54.204.85.23/wp-content/uploads/2020/12/1614778286060.jpg',
-        alt: 'yoyo'
-      }
-    }} />
-```
