--- src/elements/core-list/core-list.docs.mdx
@@ -1,35 +1,0 @@
----
-description: 'A CoreList component.'
-labels: ['label1', 'label2', 'label3']
----
-
-import { CoreList } from './core-list';
-
-## Render list from wordpress list component
-
-
-### Component usage
-```js
-<CoreList {...{
-    name: 'core/list',
-    attributes: {
-      ordered: false,
-      values:
-        '<li>Where it is located in Scripture</li><li>Who was present for this teaching</li><li>Why it was given</li><li>What is the key verse</li>'
-    }
-  }} />
-```
-
-### Using props to customize the text
-
-Modify the text to see it change live:
-```js live
-<CoreList {...{
-    name: 'core/list',
-    attributes: {
-      ordered: false,
-      values:
-        '<li>Where it is located in Scripture</li><li>Who was present for this teaching</li><li>Why it was given</li><li>What is the key verse</li>'
-    }
-  }} />
-```
