--- src/elements/core-paragraph/core-paragraph.docs.mdx
@@ -1,34 +1,0 @@
----
-description: 'A CoreParagraph component.'
-labels: ['label1', 'label2', 'label3']
----
-
-import { CoreParagraph } from './core-paragraph';
-
-## React Component for rendering text
-
-A basic div that renders some text
-
-### Component usage
-```js
-<CoreParagraph {...{
-      attributes: {
-        content:
-          "When people talk about the ministry of Jesus, it's easy to focus on his miracles. Jesus performed some amazing feats that the world had never seen (and hasn’t seen since). But one of the most exciting things about His ministry was His teaching style."
-      },
-      name: 'core/paragraph'
-    }} />
-```
-
-### Using props to customize the text
-
-Modify the text to see it change live:
-```js live
-<CoreParagraph {...{
-      attributes: {
-        content:
-          "When people talk about the ministry of Jesus, it's easy to focus on his miracles. Jesus performed some amazing feats that the world had never seen (and hasn’t seen since). But one of the most exciting things about His ministry was His teaching style."
-      },
-      name: 'core/paragraph'
-    }} />
-```
